+++ files/bash/.profile.d/bash_completion.sh
@@ -1,0 +1,2 @@
+# Configure extra shell command completions
+complete -C aws_completer aws