+++ CHANGELOG.md
@@ -1,5 +1,25 @@
 # workstation change history
 
+## 0.1.6
+
+-   Added bash-git-prompt config
+
+## 0.1.5
+
+-   Added ~/.profile.d files and updated tests
+
+## 0.1.4
+
+-   Added ~/.profile.d
+
+## 0.1.3
+
+-   Added more bash configs and updated unit/integration tests
+
+## 0.1.2
+
+-   Style fixes and test cleanup
+
 ## 0.1.1
 
 -   Added initial bash configuration